api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Mesquita Vincent <vmqt.pro@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
google-api-python-client = "^2.161.0"
fastapi = {extras = ["standard"], version = "^0.115.8"}
python-dotenv = "^1.0.1"
stringcase = "^1.2.0"
pytest = "^8.3.4"
black = "^25.1.0"
isort = "^6.0.0"
injector = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
